Added mock cell to Home
